Rename consistency to merkle
